requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
matplotlib>=3.4.0
pandas>=1.3.0
Pillow>=8.3.0
scikit-learn>=1.0.0
requests>=2.25.0
urllib3>=1.26.0
